build: switch to gulp to build package
Mainly because library packages need to be installed as a XAR and
`gulp-exist` allows that now. Also the version is now only managed in
package.json.
It is also faster and more extensible than `ant`. Building the XAR and
installing it in the database takes a little over half a second.

The project structure changed. All project files are now in `src` and
the project root contains only meta data files.